pyproject.toml
[project]
name = "test-lyria3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.68.0",
    "python-dotenv>=1.2.2",
]
